Move process_record_user to another file

+++ users/dinhhuy258/oled/dog.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void render_dog(int dog_x, int dog_y, int current_wpm);
+void render_dog(uint8_t dog_x, uint8_t dog_y, uint8_t current_wpm);
